move classes into their php own files, one class per file, class name matches file name. add psr-4 autoload.
