+++ coveopush/BatchDocument.php
@@ -1,0 +1,13 @@
+<?php
+
+// -------------------------------------------------------------------------------------
+// BatchDocument
+// -------------------------------------------------------------------------------------
+// Class to hold the Batch Document.
+// -------------------------------------------------------------------------------------
+namespace Coveo\Search\SDK\SDKPushPHP;
+
+class BatchDocument{
+    public $AddOrUpdate = array();
+    public $Delete = array();
+}
