+++ coveopush/HttpHeaders.php
@@ -1,0 +1,13 @@
+<?php
+// -------------------------------------------------------------------------------------
+// HttpHeaders
+// -------------------------------------------------------------------------------------
+// Contains the HttpHeaders Constants used by the SDK
+// -------------------------------------------------------------------------------------
+
+namespace Coveo\Search\SDK\SDKPushPHP;
+
+class HttpHeaders{
+    const AMAZON_S3_SERVER_SIDE_ENCRYPTION_NAME = "x-amz-server-side-encryption";
+    const AMAZON_S3_SERVER_SIDE_ENCRYPTION_VALUE = "AES256";
+}
