+++ coveopush/Parameters.php
@@ -1,0 +1,18 @@
+<?php
+// -------------------------------------------------------------------------------------
+// CompressionType
+// -------------------------------------------------------------------------------------
+// Contains the CompressionType used by the SDK
+// -------------------------------------------------------------------------------------
+
+namespace Coveo\Search\SDK\SDKPushPHP;
+
+class Parameters{
+    const STATUS_TYPE = "statusType";
+    const FILE_ID = "fileId";
+    const ORDERING_ID = "orderingId";
+    const DOCUMENT_ID = "documentId";
+    const QUEUE_DELAY = "queueDelay";
+    const DELETE_CHILDREN = "deleteChildren";
+    const COMPRESSION_TYPE = "compressionType";
+}
