+++ coveopush/PlatformEndpoint.php
@@ -1,0 +1,15 @@
+<?php
+// -------------------------------------------------------------------------------------
+// CoveoConstants
+// -------------------------------------------------------------------------------------
+// Contains the Constants used by the SDK
+// -------------------------------------------------------------------------------------
+
+namespace Coveo\Search\SDK\SDKPushPHP;
+
+class PlatformEndpoint{
+    const PROD_PLATFORM_API_URL = "https://platform.cloud.coveo.com";
+    const HIPAA_PLATFORM_API_URL = "https://platformhipaa.cloud.com";
+    const QA_PLATFORM_API_URL = "https://platformqa.cloud.coveo.com";
+    const DEV_PLATFORM_API_URL = "https://platformdev.cloud.coveo.com";
+}
