+++ coveopush/PlatformPaths.php
@@ -1,0 +1,12 @@
+<?php
+// -------------------------------------------------------------------------------------
+// CoveoConstants
+// -------------------------------------------------------------------------------------
+// Contains the Constants used by the SDK
+// -------------------------------------------------------------------------------------
+
+namespace Coveo\Search\SDK\SDKPushPHP;
+
+class PlatformPaths{
+    const CREATE_PROVIDER = "{endpoint}/rest/organizations/{org_id}/securityproviders/{prov_id}";
+}
