+++ coveopush/PushApiEndpoint.php
@@ -1,0 +1,15 @@
+<?php
+// -------------------------------------------------------------------------------------
+// CoveoConstants
+// -------------------------------------------------------------------------------------
+// Contains the Constants used by the SDK
+// -------------------------------------------------------------------------------------
+
+namespace Coveo\Search\SDK\SDKPushPHP;
+
+class PushApiEndpoint{
+    const PROD_PUSH_API_URL = "https://api.cloud.coveo.com/push/v1";
+    const HIPAA_PUSH_API_URL = "https://apihipaa.cloud.coveo.com/push/v1";
+    const QA_PUSH_API_URL = "https://api-qa.cloud.coveo.com/push/v1";
+    const DEV_PUSH_API_URL = "https://api-dev.cloud.coveo.com/push/v1";
+}
